Update for Stardew Valley 1.6 and SMAPI 4.0
Updated from .Net 5.0 to .Net 6.0
Updated for SMAPI 4.0 and Stardew Valley 1.6 API changes.
Added ToolData for each tool type we care about and added ToolUpdateData to each to allow for natural population of Clint's shop.
Removed Blacksmith.cs as it was not longer was needed.
Added some Harmony hooks to catch the MigrateLegacyItemId for tools already at Prismatic coming from 1.5 and set the ItemID to that which was added to the ToolData.
Removed some no longer needed Harmony hooks.
Updated some magic numbers with some predefined constants for better readability.